Refatoração V1.3
