+++ src/main/java/br/com/fiap/view/MenuContaCarteira.java
@@ -1,6 +1,9 @@
 package br.com.fiap.view;
 
+import br.com.fiap.service.CarteiraService;
 import br.com.fiap.service.ContaCarteiraService;
+import br.com.fiap.service.ContaExternaService;
+import br.com.fiap.service.ContaInternaService;
 
 import java.util.Scanner;
 
@@ -31,12 +34,12 @@ public class MenuContaCarteira {
             System.out.println();
             try {
                 switch (opcao) {
-                    case "1": ContaCarteiraService.consultarContaInterna(scanner); break;
-                    case "2": ContaCarteiraService.consultarCarteira(scanner); break;
-                    case "3": ContaCarteiraService.transferirEntreContasInternas(scanner); break;
+                    case "1": ContaInternaService.consultarContaInterna(scanner); break;
+                    case "2": CarteiraService.consultarCarteira(scanner); break;
+                    case "3": ContaInternaService.transferirEntreContasInternas(scanner); break;
                     case "4": ContaCarteiraService.adicionarSaldoConta(scanner); break;
-                    case "5": ContaCarteiraService.consultarHistoricoTransacoes(scanner); break;
-                    case "6": ContaCarteiraService.cadastrarContaExterna(scanner); break;
+                    case "5": ContaInternaService.consultarHistoricoTransacoes(scanner); break;
+                    case "6": ContaExternaService.cadastrarContaExterna(scanner); break;
                     case "0":
                         System.out.println("Voltando ao menu principal...");
                         break;
